Deal with some warnings

+++ include/global.hh
@@ -12,8 +12,9 @@
 #include <cstdint>
 #include <vector>
 #include <memory>
+#include <functional>
 typedef unsigned char uchar;        //!< The type to use for accessing individual bytes
-typedef int64_t llint;              //!< A large signed integer type.
+typedef ptrdiff_t llint;              //!< A large signed integer type.
 typedef float trace_t;              //!< Real type for traces, use "*_t" convention
 typedef double geom_t;               //!< Real type for geometry
 typedef double unit_t;               //!< Real type for physical, SI and math units
